[macOS] fix “Serialize” ambiguity in wallet header
Fixes error below:
  CXX      libbitcoin_server_a-init.o
In file included from zcash/Proof.hpp:4:0,
                 from zcash/JoinSplit.hpp:5,
                 from init.h:11,
                 from init.cpp:10:
./serialize.h: In instantiation of ‘void SerReadWrite(Stream&, const T&, int, int, CSerActionSerialize) [with Stream = CSizeComputer; T = long unsigned int]’:
wallet/wallet.h:175:9:   required from ‘void JSOutPoint::SerializationOp(Stream&, Operation, int, int) [with Stream = CSizeComputer; Operation = CSerActionSerialize]’
wallet/wallet.h:170:5:   required from here
./serialize.h:951:16: error: call of overloaded ‘Serialize(CSizeComputer&, const long unsigned int&, int&, int&)’ is ambiguous
     ::Serialize(s, obj, nType, nVersion);
     ~~~~~~~~~~~^~~~~~~~~~~~~~~~~~~~~~~~~

+++ src/wallet/wallet.h
@@ -160,7 +160,7 @@ public:
     // Transaction hash
     uint256 hash;
     // Index into CTransaction.vjoinsplit
-    size_t js;
+    uint64_t js;
     // Index into JSDescription fields of length ZC_NUM_JS_OUTPUTS
     uint8_t n;
